@@ -2,6 +2,7 @@
 import contextlib
 import hashlib
 import math
+import os
 import queue
 import random
 import struct
@@ -33,6 +34,7 @@ from vllm.v1.request import RequestStatus
 import vllm_ascend.envs as envs_ascend
 from vllm_ascend.ascend_config import get_ascend_config, init_ascend_config
 from vllm_ascend.distributed.mooncake.transfer_engine import get_global_te
+from vllm_ascend.distributed.utils import get_transfer_timeout_value
 from vllm_ascend.utils import vllm_version_is
 
 if vllm_version_is("0.11.0"):
@@ -855,6 +857,8 @@ class MooncakeConnectorWorker:
 
     def __init__(self, vllm_config: VllmConfig, engine_id: str):
         self._get_prefill_decode_size(vllm_config)
+        os.environ["ASCEND_TRANSFER_TIMEOUT"] = str(
+            get_transfer_timeout_value())
         if self._prefill_tp_size < self._decode_tp_size:
             raise ValueError(
                 f"prefill_tp_size: {self._prefill_tp_size} must be greater than"
@@ -905,7 +909,7 @@ class MooncakeConnectorWorker:
         self.device_id = device_ids[self.tp_rank]  # type: ignore
 
         if vllm_config.kv_transfer_config.get_from_extra_config(
-                'use_ascend_direct', False):
+                'use_ascend_direct', True):
             hostname = self.side_channel_host
         else:
             hostname = f"{self.side_channel_host}:0:npu_{self.device_id}"
